Finished EnzoRiemannImpl refactor to use kernel functors
