+++ src/Enzo/EnzoRiemann/EnzoRiemannUtils.hpp
@@ -260,87 +260,6 @@ namespace enzo_riemann_utils{
 
   //----------------------------------------------------------------------
 
-  /// Constructs an array of instances of EFlt3DArray this is something of a
-  /// stopgap solution until we fully support implemntation of EFlt3DArray
-  /// with 4D CelloArrays
-  template<class LUT>
-  inline std::array<CelloArray<const enzo_float, 3>, LUT::num_entries>
-  array_from_map(const EnzoEFltArrayMap& map) noexcept
-  {
-    std::array<CelloArray<const enzo_float, 3>, LUT::num_entries> out;
-    for (std::size_t i = 0; i < LUT::num_entries; i++) { out[i] = map[i]; }
-    return out;
-  }
-
-  template<class LUT>
-  inline std::array<CelloArray<enzo_float, 3>,LUT::num_entries> array_from_map
-  (EnzoEFltArrayMap& map) noexcept
-  {
-    std::array<CelloArray<enzo_float, 3>, LUT::num_entries> out;
-    for (std::size_t i = 0; i < LUT::num_entries; i++) { out[i] = map[i]; }
-    return out;
-  }
-
-  //----------------------------------------------------------------------
-
-  /// @def      LUT_TRANSFER_T_SCALAR
-  /// @brief    Part of the LUT_TRANSFER group of macros that are used to copy
-  ///           scalar actively advected values between a
-  ///           std::array<enzo_float, LUT::num_entries> and a given
-  ///           location in a list of 3D external arrays. These macros simply
-  ///           ignore the vector quantities (since that involves some level of
-  ///           permutation)
-  ///
-  /// The macro copies values from `src` to `dest`. When `src` (`dest`) is the
-  /// list of external arrays, `SRCSUF` (`DESTSUF`) should be the parentheses
-  /// enclosed indices of the arrays while `DESTSUF` (`SRCSUF`) should be passed
-  /// an empty argument.
-  #define LUT_TRANSFER_T_SCALAR(LUT, src, dest, SRCSUF, DESTSUF, name)        \
-    if (LUT::name != -1){ dest[LUT::name] DESTSUF = src[LUT::name] SRCSUF; }
-  #define LUT_TRANSFER_T_VECTOR(LUT, src, dest, SRCSUF, DESTSUF, name) /*...*/
-  #define LUT_TRANSFER_F_SCALAR(LUT, src, dest, SRCSUF, DESTSUF, name) /*...*/
-  #define LUT_TRANSFER_F_VECTOR(LUT, src, dest, SRCSUF, DESTSUF, name) /*...*/
-
-  //----------------------------------------------------------------------
-
-  /// transfers actively advected scalar quantities values from the specified
-  /// correct location in `external` to a 1D output array.
-  template<class LUT>
-  void transfer_scalars_to_lutarray
-  (const int dim, const int iz, const int iy, const int ix,
-   const std::array<CelloArray<const enzo_float, 3>,LUT::num_entries>& external,
-   lutarray<LUT> &dest)
-    noexcept
-  {
-    #define ENTRY(name, math_type, category, if_advection)                    \
-      LUT_TRANSFER_##if_advection##_##math_type (EnzoRiemannLUT<LUT>,         \
-                                                 external, dest, (iz,iy,ix), ,\
-                                                 name);
-    FIELD_TABLE
-    #undef ENTRY
-  }
-
-  //----------------------------------------------------------------------
-
-  /// transfers actively advected scalar quantities values from `src` to the
-  /// correct location in `external`.
-  template<class LUT>
-  inline void transfer_scalars_from_lutarray
-  (const int dim, const int iz, const int iy, const int ix,
-   const std::array<CelloArray<enzo_float, 3>, LUT::num_entries>& external,
-   const lutarray<LUT> src) noexcept
-  {
-    #define ENTRY(name, math_type, category, if_advection)                    \
-      LUT_TRANSFER_##if_advection##_##math_type (EnzoRiemannLUT<LUT>,         \
-                                                 src, external, , (iz,iy,ix), \
-                                                 name);
-    FIELD_TABLE
-    #undef ENTRY
-  }
-
-
-  //----------------------------------------------------------------------
-
   /// computes the flux of a passive scalar at a given cell interface.
   ///
   /// @param left,right The passive scalars (as mass fractions) on the left and
